requirements.txt
anthropic==0.49.0
pdf2image==1.16.3
pillow==10.1.0
numpy==1.26.2
streamlit==1.31.0
transformers==4.36.2
torch==2.1.1

